Rename required files
