+++ include/cloysterhpc/services/repo.h
@@ -58,72 +58,4 @@ public:
     }
 };
 
-/**
- * Generic file class
- *
- * This class should only read and write,
- */
-class GenericFile {
-protected:
-    std::filesystem::path m_path;
-
-public:
-    explicit GenericFile(const std::filesystem::path& path)
-        : m_path(path)
-    {
-    }
-
-    virtual void read() { }
-    virtual void write() { }
-
-    virtual ~GenericFile() = default;
-};
-
-struct ELCloneRepo {
-    std::string group;
-    std::string name;
-    std::optional<std::string> base_url;
-    std::optional<std::string> metalink;
-    bool enabled;
-    bool gpgcheck;
-    std::string gpgkey;
-
-    // (P/ todos os repositórios)
-    // std::string release;
-};
-
-/**
- * Repository file class
- *
- * This class should parse the repository data
- */
-class ELRepoFile : GenericFile {
-private:
-    Glib::RefPtr<Glib::KeyFile> m_file;
-
-    std::vector<ELCloneRepo> parseData();
-    void unparseData(const std::vector<ELCloneRepo>& data);
-
-    std::vector<ELCloneRepo> m_repositories;
-
-public:
-    explicit ELRepoFile(const std::filesystem::path& path)
-        : GenericFile(path)
-    {
-    }
-
-    virtual void read() override;
-    virtual void write() override;
-
-    void parse();
-    void parse(const std::stringstream& ss);
-    void unparse();
-    void unparse(std::stringstream& ss);
-
-    [[nodiscard]] std::vector<ELCloneRepo>& getRepositories();
-    [[nodiscard]] const std::vector<ELCloneRepo>& getRepositoriesConst() const;
-
-    ~ELRepoFile() override = default;
-};
-
 #endif // CLOYSTERHPC_REPO_H_
